Added the creation and use of pickle files
Using the pickle module it is possible to save a classifier to be used later. On this commit the creation of such file for both SVR and linear regression is done..
Furthermore the loading and implementation of the linear regression pickle file is done, saving resources from the training stage on the script.